backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pdfplumber>=0.11.0
motor>=3.4.0
pymongo[srv]>=4.7.0
